requirements.txt
xarray
numpy
matplotlib
pillow
geojson
aiohttp
netCDF4
scipy
